Add pep8 styling
