+++ api/controllers/group.py
@@ -238,58 +238,4 @@ class GroupController():
             )
         return response
 
-    def group_message(self, data, grp_id):
-        if not request.data:
-            return (
-                jsonify(
-                    {
-                        "error": "Please provide details",
-                        "status": 400,
-                    }
-                ),
-                400,
-            )
-        data = request.get_json(force=True)
-
-        group_message_data = {
-            "subject": data.get("subject"),
-            "message": data.get("message"),
-            "sender_status": "sent",
-            "receiver_status": "unread",
-            "group_id": data.get("groupId"),
-        }
-
-        not_valid = validate_new_message(**group_message_data)
-        known_group = db.get_group(data.get("groupId"))
-        response = None
-        if not known_group:
-            response = (
-                jsonify({
-                    "status": 404, 
-                    "error": "Group with such Id does not exist."}
-                ),
-                404,
-            )  
-        elif not_valid:
-            response = not_valid 
-           
-        else:
-            group_message_data["user_id"] = get_current_identity()["email"]
-            new_message = db.create_group_message(**group_message_data)
-
-            response = (
-                jsonify(
-                    {
-                        "status": 201,
-                        "data": [
-                            {
-                                "mail": new_message,
-                                "message": "Sent message successfully",
-                            }
-                        ],
-                    }
-                ),
-                201,
-            )
-        return response
-   
+
